requirements.txt
chromadb>=0.5.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.27.0
PyMuPDF>=1.24.0
markdown>=3.5.0
rich>=13.7.0
typer>=0.9.0
jinja2>=3.1.0
python-multipart>=0.0.9
beautifulsoup4>=4.12.0
python-docx>=1.1.0
